refactor(OIDC): Give each trust relationship component its own folder
Adopt the component-folder layout the frontend guideline expects: every
new component now lives in its own directory with an index.ts barrel
(ComponentName/ComponentName.tsx + index.ts). providers.ts and errors.ts
stay as flat modules, being exempt constants/util. Cross-component
imports use the components/ alias, as the import-alias lint rule requires
for parent-traversal paths; APIKeysTab keeps importing the tab via the
unchanged ./trust-relationships path.

beep boop
