+++ frontend/web/components/pages/organisation-settings/tabs/trust-relationships/NewTrustRelationshipModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NewTrustRelationshipModal'
