+++ frontend/web/components/pages/organisation-settings/tabs/trust-relationships/TrustRelationshipModal/TrustRelationshipModal.tsx
@@ -19,8 +19,10 @@ import {
 } from 'common/services/useMasterAPIKeyWithMasterAPIKeyRole'
 import TrustRelationshipPermissionsFields, {
   SelectedRole,
-} from './TrustRelationshipPermissionsFields'
-import WorkflowSetupSnippet, { GITHUB_ISSUER } from './WorkflowSetupSnippet'
+} from 'components/pages/organisation-settings/tabs/trust-relationships/TrustRelationshipPermissionsFields'
+import WorkflowSetupSnippet, {
+  GITHUB_ISSUER,
+} from 'components/pages/organisation-settings/tabs/trust-relationships/WorkflowSetupSnippet'
 
 type ClaimRuleRow = { claim: string; values: string }
