+++ frontend/web/components/pages/organisation-settings/tabs/trust-relationships/TrustRelationshipPermissionsFields/index.ts
@@ -1,0 +1,2 @@
+export { default } from './TrustRelationshipPermissionsFields'
+export type { SelectedRole } from './TrustRelationshipPermissionsFields'
