+++ frontend/web/components/pages/organisation-settings/tabs/trust-relationships/TrustRelationships/TrustRelationships.tsx
@@ -6,12 +6,12 @@ import Icon from 'components/icons/Icon'
 import PageTitle from 'components/PageTitle'
 import PanelSearch from 'components/PanelSearch'
 import Switch from 'components/Switch'
-import GithubTrustRelationshipForm from './GithubTrustRelationshipForm'
-import { trustRelationshipErrorMessage } from './errors'
-import NewTrustRelationshipModal from './NewTrustRelationshipModal'
-import TrustRelationshipModal from './TrustRelationshipModal'
-import { isGithubFormEditable } from './WorkflowSetupSnippet'
-import { providerForIssuer } from './providers'
+import GithubTrustRelationshipForm from 'components/pages/organisation-settings/tabs/trust-relationships/GithubTrustRelationshipForm'
+import { trustRelationshipErrorMessage } from 'components/pages/organisation-settings/tabs/trust-relationships/errors'
+import NewTrustRelationshipModal from 'components/pages/organisation-settings/tabs/trust-relationships/NewTrustRelationshipModal'
+import TrustRelationshipModal from 'components/pages/organisation-settings/tabs/trust-relationships/TrustRelationshipModal'
+import { isGithubFormEditable } from 'components/pages/organisation-settings/tabs/trust-relationships/WorkflowSetupSnippet'
+import { providerForIssuer } from 'components/pages/organisation-settings/tabs/trust-relationships/providers'
 import { TrustRelationship } from 'common/types/responses'
 import {
   useDeleteTrustRelationshipMutation,
