+++ frontend/web/components/pages/organisation-settings/tabs/trust-relationships/TrustRelationships/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TrustRelationships'
