+++ frontend/web/components/pages/organisation-settings/tabs/trust-relationships/WorkflowSetupSnippet/index.ts
@@ -1,0 +1,6 @@
+export {
+  default,
+  GITHUB_ISSUER,
+  isGithubFormEditable,
+  isDefaultGithubAudience,
+} from './WorkflowSetupSnippet'
